+++ src/controllers/OrderController.php
@@ -13,6 +13,7 @@ use craft\commerce\Plugin as Commerce;
 use craft\errors\MissingComponentException;
 use craft\web\Controller;
 use ether\mc\MailchimpCommerce;
+use Exception;
 use yii\web\BadRequestHttpException;
 use yii\web\Response;
 
@@ -41,6 +42,7 @@ class OrderController extends Controller
 		$session = Craft::$app->getSession();
 
 		$number = Craft::$app->getRequest()->getRequiredQueryParam('number');
+		$cid = Craft::$app->getRequest()->getQueryParam('mc_cid');
 		$order = $commerce->getOrders()->getOrderByNumber($number);
 
 		if (!$order)
@@ -53,6 +55,25 @@ class OrderController extends Controller
 		{
 			$session->setError($settings->completedCartError);
 			return $this->redirect($settings->abandonedCartRestoreUrl);
+		}
+
+		if ($cid)
+		{
+			try
+			{
+				Craft::$app->getDb()->createCommand()
+					->update(
+						'{{%mc_orders_synced}}',
+						['cid' => $cid],
+						['orderId' => $order->id],
+						[],
+						false
+					)->execute();
+			}
+			catch (Exception $e)
+			{
+				Craft::error($e, 'mailchimp-commerce');
+			}
 		}
 
 		$commerce->getCarts()->forgetCart();