+++ CHANGELOG.md
@@ -1,6 +1,10 @@
+## [Unreleased] 1.0.8
+### Added
+- Restored carts now track their email campaign ID and sync it back to the order
+
 ## 1.0.7 - 2019-06-25
 ### Fixed
-- Fix incorrect request method error when updating an order 
+- Fix incorrect request method error when updating an order
 
 ## 1.0.6 - 2019-06-25
 ### Fixed